move anomaly function to utils
